openstack: Only disable Swift on container listing on auth error
On first boot, cluster-image-registry-operator on OpenStack tries to
detect whether object storage is available as a backend.

If it obtains access to the object storage endpoint, CIRO does one last
check: it tries to list the existing containers. Before this patch, a
failure to list containers resulted in falling over to Cinder volumes.
With this patch, CIRO will fall over to Cinder volumes if and only if
Swift returns authentication (401) or authorization (403) error codes.
Alternatively the check is retried.